restructure and enhance ComboBox validation tests
